+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Documentation - Switch to pydata sphinx theme on readthedocs (!479)
 - Data - The `long_name` and `name` attributes of dataset coordinates now contain information about the root instrument and submodules in addition to the settable (!478)
 - SI utilities - Handle nan in value_precision input (!487)
+- Utilities - Create utility function to compare snapshots (!485)
 
 ## 0.7.3 (2023-08-17)